Conflicts:
	cocos2dx/label_nodes/CCFontAtlasFactory.cpp
	samples/Cpp/TestCpp/Classes/LabelTest/LabelTest.cpp

+++ cocos2dx/include/ccTypeInfo.h
@@ -41,7 +41,7 @@ public:
 
 static inline unsigned int getHashCodeByString(const char *key)
 {
-	unsigned int len = strlen(key);
+	size_t len = strlen(key);
 	const char *end=key+len;
 	unsigned int hash;
